Remove jackson-databind dependency
Github has issued a notification about ocb-tools using a version of
jackson-databind which has a security vulnerability. It turns out that
dependency is not actually used for anything so the solution is to
simply remove it.

After rebuild ocb-tools would fail because of changed interface in
rawrepo dependency. Version 1.5 is ancient so the version was updated to
current version and that solved the problem.