Clean cache table between test runs
We need to clean the record cache table after a test case. If update
returns a OK reply the cache table will most likely be empty, but if
update returns a failure and commits, then there will be an echo left
in the cache which might cause trouble for the following test case.

Records summary was added after the last update to ocb-tools and should
be cleared between runs so I'll just add it to this change.