Use holdings-items-postgres-1.1.4
The old 1.1.1 database was hopelessly deprecated so we have to use a
newer version.

The new version uses flyway to create the database. In order to fit that
into the existing code we perform the migration on every test run. This
is not efficient and should be optimized "at some point".

Flyway spams a lot of log and since we don't want any log we need to
explicitly suppress all that logging.